рефакторинг EventPresenter и EventFragment
