+++ app/src/main/java/ru/myproevent/domain/di/ItemPresentersModule.kt
@@ -1,0 +1,4 @@
+package ru.myproevent.domain.di
+
+class ItemPresentersModule {
+}
