+++ app/src/main/java/ru/myproevent/ui/adapters/event_items/presenter_interfaces/IEventDateItemPresenter.kt
@@ -1,9 +1,9 @@
 package ru.myproevent.ui.adapters.event_items.presenter_interfaces
 
-import ru.myproevent.ui.adapters.IItemPresenter
 import ru.myproevent.ui.adapters.event_items.view_item_interfaces.IEventDateItemView
+import ru.myproevent.ui.presenters.events.event.event_screen_items_presenters.ItemPresenter
 
-interface IEventDateItemPresenter: IItemPresenter<IEventDateItemView> {
-    fun onEditClick(view: IEventDateItemView)
-    fun onRemoveClick(view: IEventDateItemView)
+abstract class IEventDateItemPresenter: ItemPresenter<IEventDateItemView>() {
+    abstract fun onEditClick(view: IEventDateItemView)
+    abstract fun onRemoveClick(view: IEventDateItemView)
 }
