+++ app/src/main/java/ru/myproevent/ui/adapters/event_items/presenter_interfaces/IEventScreenListPresenter.kt
@@ -4,9 +4,9 @@ import ru.myproevent.ui.adapters.IItemPresenter
 import ru.myproevent.ui.adapters.event_items.view_item_interfaces.*
 
 interface IEventScreenListPresenter {
-    val proeventProfilePictureFormItemPresenter: IItemPresenter<IProeventProfilePictureItemView>
-    val proeventTextFormItemPresenter: IItemPresenter<IProeventTextFormItemView>
-    val proeventFormsHeaderItemPresenter: IFormsHeaderItemPresenter
+    val profilePictureFormItemPresenter: IItemPresenter<IProeventProfilePictureItemView>
+    val textFormItemPresenter: IItemPresenter<IProeventTextFormItemView>
+    val formsHeaderItemPresenter: IFormsHeaderItemPresenter
     val participantItemPresenter: IItemPresenter<IParticipantItemView>
     val eventDateItemPresenter: IEventDateItemPresenter
     val noItemsPlaceholderItemPresenter: IItemPresenter<INoItemsPlaceholderItemView>
