+++ app/src/main/java/ru/myproevent/ui/data_items/no_items_placeholders/NoDatesPlaceholderFactory.kt
@@ -1,0 +1,16 @@
+package ru.myproevent.ui.data_items.no_items_placeholders
+
+import ru.myproevent.R
+import ru.myproevent.ui.adapters.event_items.EventScreenItem
+
+class NoDatesPlaceholderFactory {
+    companion object {
+        fun create(header: EventScreenItem.FormsHeader<EventScreenItem.ListItem>) =
+            EventScreenItem.NoItemsPlaceholder(
+                "Отсутствует.\nНажмите + чтобы добавить.",
+                R.drawable.ic_add,
+                21,
+                header
+            )
+    }
+}
