+++ app/src/main/java/ru/myproevent/ui/presenters/events/event/EventPresenter.kt
@@ -1,6 +1,5 @@
 package ru.myproevent.ui.presenters.events.event
 
-import android.util.Log
 import android.widget.Toast
 import com.github.terrakok.cicerone.Router
 import ru.myproevent.ProEventApp
@@ -13,13 +12,10 @@ import ru.myproevent.domain.models.repositories.events.IProEventEventsRepository
 import ru.myproevent.domain.models.repositories.images.IImagesRepository
 import ru.myproevent.domain.models.repositories.proevent_login.IProEventLoginRepository
 import ru.myproevent.domain.models.repositories.profiles.IProEventProfilesRepository
-import ru.myproevent.domain.utils.toParticipantItems
 import ru.myproevent.ui.adapters.event_items.*
-import ru.myproevent.ui.adapters.event_items.presenter_interfaces.IEventScreenListPresenter
-import ru.myproevent.ui.adapters.event_items.presenter_interfaces.IFormsHeaderItemPresenter
-import ru.myproevent.ui.adapters.IItemPresenter
-import ru.myproevent.ui.adapters.event_items.presenter_interfaces.IEventDateItemPresenter
-import ru.myproevent.ui.adapters.event_items.view_item_interfaces.*
+import ru.myproevent.ui.data_items.EventScreenItemsFactory
+import ru.myproevent.ui.data_items.no_items_placeholders.NoDatesPlaceholderFactory
+import ru.myproevent.ui.data_items.no_items_placeholders.NoItemsPlaceholderFactory
 import ru.myproevent.ui.presenters.BaseMvpPresenter
 import java.io.File
 import java.lang.RuntimeException
@@ -41,604 +37,17 @@ class EventPresenter(localRouter: Router, var eventBeforeEdit: Event?) :
     @Inject
     lateinit var imagesRepository: IImagesRepository
 
-    val participantProfiles = mutableMapOf<Long, Profile>()
-
-//    private val dates: MutableList<TimeInterval>? = mutableListOf(
-//        TimeInterval(1646817780000, 1646821380000),
-//        TimeInterval(1643977614L * 1000L, (1643977614L + 3600L) * 1000L),
-//        TimeInterval((1646815896L - 3600L * 2L) * 1000L, 1646815896L * 1000L),
-//        TimeInterval(1646815896L * 1000L, (1646815896L + 3600L * 2L) * 1000L),
-//        TimeInterval((1646815896L + 86400L * 2L) * 1000L, (1646815896L + 86400L * 3L) * 1000L)
-//    ).apply {
-//        sortWith { a, b ->
-//            val longDiff = a.start - b.start
-//            return@sortWith if (longDiff > 0L) {
-//                1
-//            } else {
-//                -1
-//            }
-//        }
-//    }
-
-    private fun removeHeaderItemsFromScreenItems(headerPosition: Int) {
-        eventScreenListPresenter.eventScreenItems.subList(
-            headerPosition + 1,
-            headerPosition + 1 + (eventScreenListPresenter.eventScreenItems[headerPosition] as EventScreenItem.FormsHeader<EventScreenItem.ListItem>).items.size
-        ).clear()
-    }
-
-    private fun addHeaderItemsToScreenItems(
-        headerPosition: Int,
-        items: TreeSet<EventScreenItem.ListItem>
-    ) {
-        eventScreenListPresenter.eventScreenItems.addAll(headerPosition + 1, items)
-    }
-
-    private fun enableDescriptionEdit(headerPosition: Int) {
-        with(eventScreenListPresenter.eventScreenItems[headerPosition] as EventScreenItem.FormsHeader<EventScreenItem.ListItem>) {
-            if (items.first() is EventScreenItem.NoItemsPlaceholder) {
-                removeHeaderItemsFromScreenItems(headerPosition)
-                addHeaderItemsToScreenItems(headerPosition, items.apply {
-                    clear()
-                    add(
-                        EventScreenItem.TextBox(
-                            value = "",
-                            header = this@with,
-                            isEditLocked = false,
-                            hasFocusIntent = true
-                        )
-                    )
-                })
-                viewState.updateEventScreenList()
-            } else {
-                if (!isExpanded) {
-                    isExpanded = true
-                    addHeaderItemsToScreenItems(headerPosition, items)
-                }
-                with(items.first() as EventScreenItem.TextBox) {
-                    isEditLocked = false
-                    hasFocusIntent = true
-                }
-                viewState.updateEventScreenList()
-            }
-        }
-    }
-
-    private class NoDescriptionPlaceholderFactory {
-        companion object {
-            fun create(header: EventScreenItem.FormsHeader<EventScreenItem.ListItem>) =
-                EventScreenItem.NoItemsPlaceholder(
-                    "Отсутствует.\nНажмите + чтобы добавить.",
-                    R.drawable.ic_edit_blue,
-                    21,
-                    header
-                )
-
-        }
-    }
-
-    private class NoDatesPlaceholderFactory {
-        companion object {
-            fun create(header: EventScreenItem.FormsHeader<EventScreenItem.ListItem>) =
-                EventScreenItem.NoItemsPlaceholder(
-                    "Отсутствует.\nНажмите + чтобы добавить.",
-                    R.drawable.ic_add,
-                    21,
-                    header
-                )
-        }
-    }
-
-    private class NoItemsPlaceholderFactory {
-        companion object {
-            fun create(header: EventScreenItem.FormsHeader<EventScreenItem.ListItem>) =
-                EventScreenItem.NoItemsPlaceholder(
-                    "Отсутствуют.\nНажмите + чтобы добавить.",
-                    R.drawable.ic_add,
-                    21,
-                    header
-                )
-        }
-    }
-
-    private val setOfExpandedItems = mutableSetOf<EVENT_SCREEN_ITEM_ID>()
-
-    private val isCurrentUserOwnsEvent by lazy {
+    val isCurrentUserOwnsEvent by lazy {
         eventBeforeEdit?.let { loginRepository.getLocalId()!! == it.ownerUserId } ?: true
     }
 
-    private fun getEventScreenItemsBeforeEdit(): MutableList<EventScreenItem> {
-        return mutableListOf(
-            EventScreenItem.ProfileImageForm(
-                itemId = EVENT_SCREEN_ITEM_ID.EVENT_PICTURE,
-                text = "Изменить фото мероприятия"
-            ),
-            EventScreenItem.TextForm(
-                itemId = EVENT_SCREEN_ITEM_ID.EVENT_NAME,
-                title = "Название",
-                hint = "Введите название",
-                value = eventBeforeEdit?.name ?: "",
-                isEditLocked = true,
-                isEditOptionAvailable = isCurrentUserOwnsEvent
-            ),
-            EventScreenItem.TextForm(
-                itemId = EVENT_SCREEN_ITEM_ID.LOCATION,
-                title = "Место проведения",
-                hint = "Введите адрес",
-                value = eventBeforeEdit?.address?.toString() ?: "",
-                isEditLocked = true,
-                isEditOptionAvailable = isCurrentUserOwnsEvent
-            ),
-            EventScreenItem.FormsHeader(
-                itemId = EVENT_SCREEN_ITEM_ID.DATES_HEADER,
-                title = "Время проведения",
-                isExpanded = setOfExpandedItems.contains(EVENT_SCREEN_ITEM_ID.DATES_HEADER),
-                items = TreeSet(), // TODO: отрефакторить: как избежать создание этого пустого mutableList? Он нужен просто как загулшка
-                editOptionIcon = if (isCurrentUserOwnsEvent) R.drawable.ic_add else null,
-                onEditOptionClick = {
-                    openDatePicker(null)
-                }
-            ).apply {
-                if (!eventBeforeEdit?.dates.isNullOrEmpty()) {
-                    eventBeforeEdit!!.dates.forEach { timeInterval ->
-                        items.add(EventScreenItem.EventDateItem(timeInterval, this))
-                    }
-                } else {
-                    items.add(
-                        NoDatesPlaceholderFactory.create(header = this)
-                    )
-                }
-            },
-            EventScreenItem.FormsHeader(
-                itemId = EVENT_SCREEN_ITEM_ID.DESCRIPTION_HEADER,
-                title = "Описание",
-                isExpanded = setOfExpandedItems.contains(EVENT_SCREEN_ITEM_ID.DESCRIPTION_HEADER),
-                items = TreeSet(),
-                editOptionIcon = if (isCurrentUserOwnsEvent) R.drawable.ic_edit_blue else null,
-                onEditOptionClick = { }
-            ).apply {
-                items.add(
-                    if (eventBeforeEdit != null && !eventBeforeEdit!!.description.isNullOrBlank()) {
-                        EventScreenItem.TextBox(
-                            value = eventBeforeEdit!!.description!!,
-                            header = this,
-                            isEditLocked = true,
-                            hasFocusIntent = false
-                        )
-                    } else {
-                        NoDescriptionPlaceholderFactory.create(header = this)
-                    }
-                )
+    fun getEventScreenItemsBeforeEdit() = EventScreenItemsFactory.create(eventPresenter = this)
 
-                onEditOptionClick = { headerPosition ->
-                    editOptionIcon = null
-                    enableDescriptionEdit(headerPosition)
-                    eventScreenListPresenter.proeventFormsHeaderItemPresenter.absoluteFormsHeaderPresenter.updateAbsoluteFormsHeader()
-                    viewState.showEditOptions()
-                }
-            },
-            EventScreenItem.FormsHeader(
-                itemId = EVENT_SCREEN_ITEM_ID.MAPS_HEADER,
-                title = "Карты меропрития",
-                isExpanded = setOfExpandedItems.contains(EVENT_SCREEN_ITEM_ID.MAPS_HEADER),
-                items = TreeSet(),
-                editOptionIcon = if (isCurrentUserOwnsEvent) R.drawable.ic_add else null,
-                onEditOptionClick = {
-                    viewState.showMessage("Данный функционал пока не реализован, так как для него нет законченного дизайна")
-                }
-            ).apply {
-                items.add(NoItemsPlaceholderFactory.create(header = this))
-            },
-            EventScreenItem.FormsHeader(
-                itemId = EVENT_SCREEN_ITEM_ID.POINTS_HEADER,
-                title = "Точки",
-                isExpanded = setOfExpandedItems.contains(EVENT_SCREEN_ITEM_ID.POINTS_HEADER),
-                items = TreeSet(),
-                editOptionIcon = if (isCurrentUserOwnsEvent) R.drawable.ic_add else null,
-                onEditOptionClick = {
-                    viewState.showMessage("Данный функционал пока не реализован, так как для него нет законченного дизайна")
-                }
-            ).apply {
-                items.add(NoItemsPlaceholderFactory.create(header = this))
-            },
-            EventScreenItem.FormsHeader(
-                itemId = EVENT_SCREEN_ITEM_ID.PARTICIPANTS_HEADER,
-                title = "Участники",
-                isExpanded = setOfExpandedItems.contains(EVENT_SCREEN_ITEM_ID.PARTICIPANTS_HEADER),
-                items = TreeSet(),
-                editOptionIcon = if (isCurrentUserOwnsEvent) R.drawable.ic_add else null,
-                onEditOptionClick = {
-                    localRouter.navigateTo(
-                        screens.participantPickerTypeSelection(
-                            pickedParticipantsIds
-                        )
-                    )
-                }
-            ).apply {
-                val participantItems =
-                    eventBeforeEdit?.participantsUserIds?.toParticipantItems(this)
+    val eventScreenListPresenter = EventScreenListPresenter(eventPresenter = this)
 
-                if (!participantItems.isNullOrEmpty()) {
-                    participantItems.forEach {
-                        items.add(it)
-                    }
-                } else {
-                    items.add(
-                        NoItemsPlaceholderFactory.create(header = this)
-                    )
-                }
-            }
-        )
-    }
+    val eventEditOptionsPresenter = EventEditOptionsPresenter(eventPresenter = this)
 
-    inner class EventScreenListPresenter(
-        val eventScreenItems: MutableList<EventScreenItem> = mutableListOf(),
-        private var proeventTextFormItemClickListener: (
-            (
-            IProeventTextFormItemView,
-            EventScreenItem
-        ) -> Unit)? = null,
-        private var proeventFormsHeaderItemClickListener: ((IProeventFormsHeaderItemView, EventScreenItem) -> Unit)? =
-            null,
-        override val proeventProfilePictureFormItemPresenter: IItemPresenter<IProeventProfilePictureItemView> =
-            object :
-                IItemPresenter<IProeventProfilePictureItemView> {
-                override fun onItemClick(view: IProeventProfilePictureItemView) {
-                    // TODO: stuff
-                }
-
-                override fun bindView(view: IProeventProfilePictureItemView) {
-                    val pos = view.pos
-                    with(view) {
-                        with(eventScreenItems[pos] as EventScreenItem.ProfileImageForm) {
-                            setText(text)
-                        }
-                    }
-                }
-            },
-        override val proeventTextFormItemPresenter: IItemPresenter<IProeventTextFormItemView> = object :
-            IItemPresenter<IProeventTextFormItemView> {
-            override fun onItemClick(view: IProeventTextFormItemView) {
-                proeventTextFormItemClickListener?.invoke(view, eventScreenItems[view.pos])
-            }
-
-            override fun bindView(view: IProeventTextFormItemView) {
-                val pos = view.pos
-                with(view) {
-                    with(eventScreenItems[pos] as EventScreenItem.TextForm) {
-                        setTitle(title)
-                        setHint(hint)
-                        setValue(value)
-                        Log.d(
-                            "[MYLOG]",
-                            "pos: $pos; setValue: $value; eventScreenItems[pos].itemId: ${eventScreenItems[pos].itemId}"
-                        )
-                        setEditOption(isEditOptionAvailable)
-                        setEditLock(isEditLocked)
-                        setOnEditUnlockListener {
-                            isEditLocked = false
-                            viewState.showEditOptions()
-                        }
-                        setOnEditOptionHideListener {
-                            isEditOptionAvailable = false
-                        }
-                        setOnValueChangedListener { newValue ->
-                            Log.d(
-                                "[MYLOG]",
-                                "eventScreenItems[pos].itemId: ${eventScreenItems[pos].itemId} setOnValueChangedListener newValue: $newValue"
-                            )
-                            value = newValue
-                        }
-                    }
-                }
-            }
-        },
-        override val proeventFormsHeaderItemPresenter: IFormsHeaderItemPresenter =
-            object :
-                IFormsHeaderItemPresenter {
-
-                private fun onExpandEvent(headerPosition: Int) {
-                    viewState.hideKeyboard()
-                    with(eventScreenItems[headerPosition] as EventScreenItem.FormsHeader<EventScreenItem.ListItem>) {
-                        if (!isExpanded) {
-                            addHeaderItemsToScreenItems(headerPosition, items)
-                            setOfExpandedItems.add(eventScreenItems[headerPosition].itemId)
-                            //viewState.eventScreenListNotifyItemRangeInserted(headerPosition + 1, items.size)
-                        } else {
-                            removeHeaderItemsFromScreenItems(headerPosition)
-                            setOfExpandedItems.remove(eventScreenItems[headerPosition].itemId)
-                            //viewState.eventScreenListNotifyItemRangeRemoved(headerPosition + 1, items.size)
-                        }
-                        this.isExpanded = !this.isExpanded
-                    }
-                    viewState.updateEventScreenList()
-                }
-
-                override val absoluteFormsHeaderPresenter =
-                    object : IFormsHeaderItemPresenter.AbsoluteFormsHeaderPresenter() {
-                        override fun showAbsoluteFormsHeader(currAbsoluteFormsHeaderId: EVENT_SCREEN_ITEM_ID) {
-                            super.showAbsoluteFormsHeader(currAbsoluteFormsHeaderId)
-                            val header =
-                                eventScreenItems.find { item -> item.itemId == currAbsoluteFormsHeaderId } as EventScreenItem.FormsHeader<EventScreenItem.ListItem>
-                            viewState.showAbsoluteFormsHeader(title = header.title,
-                                editIcon = header.editOptionIcon,
-                                editIconTint = null,
-                                onCollapse = {
-                                    onExpandEvent(headerPosition = eventScreenListPresenter.eventScreenItems.indexOfFirst { it == header })
-                                    eventScreenListPresenter.proeventFormsHeaderItemPresenter.absoluteFormsHeaderPresenter.hideAbsoluteFormsHeader()
-                                },
-                                onCollapseScrollToPosition = eventScreenItems.indexOfFirst { it == header },
-                                // TODO: отрефакторить: избавиться от использования headerPosition и передавать вместо этого прямую сслку на header
-                                onEdit = { header.onEditOptionClick(eventScreenItems.indexOfFirst { item -> item.itemId == header.itemId }) })
-                        }
-
-                        override fun hideAbsoluteFormsHeader() {
-                            super.hideAbsoluteFormsHeader()
-                            viewState.hideAbsoluteBar()
-                        }
-                    }
-
-                override fun onFirstVisibleItemPositionChangeListener(position: Int) {
-                    with(eventScreenItems[position]) {
-                        if (this is EventScreenItem.FormsHeader<*> && this.isExpanded) {
-                            absoluteFormsHeaderPresenter.showAbsoluteFormsHeader(this.itemId)
-                        } else if (this is EventScreenItem.ListItem && eventScreenItems[position + 1] is EventScreenItem.ListItem) {
-                            absoluteFormsHeaderPresenter.showAbsoluteFormsHeader(this.header.itemId)
-                        } else {
-                            absoluteFormsHeaderPresenter.hideAbsoluteFormsHeader()
-                        }
-                    }
-                }
-
-                override fun onEditClick(view: IProeventFormsHeaderItemView) {
-                    (eventScreenItems[view.pos] as EventScreenItem.FormsHeader<*>).onEditOptionClick(
-                        view.pos
-                    )
-                }
-
-                override fun onItemClick(view: IProeventFormsHeaderItemView) =
-                    onExpandEvent(headerPosition = eventScreenListPresenter.eventScreenItems.indexOfFirst { item ->
-                        if (item is EventScreenItem.FormsHeader<*>) {
-                            item.itemId == view.itemId
-                        } else {
-                            false
-                        }
-                    })
-
-                override fun bindView(view: IProeventFormsHeaderItemView) {
-                    val pos = view.pos
-                    with(eventScreenItems[pos] as EventScreenItem.FormsHeader<*>) {
-                        view.itemId = itemId
-                        view.setTitle(title)
-                        view.setExpandState(isExpanded)
-                        view.setEditOptionIcon(editOptionIcon)
-                    }
-                }
-            },
-        override val participantItemPresenter: IItemPresenter<IParticipantItemView> = object :
-            IItemPresenter<IParticipantItemView> {
-            override fun onItemClick(view: IParticipantItemView) {
-                val profile =
-                    participantProfiles[(eventScreenItems[view.pos] as EventScreenItem.ParticipantItem).participantId]!!
-                localRouter.navigateTo(screens.eventParticipant(profile))
-            }
-
-            override fun bindView(view: IParticipantItemView) {
-                val pos = view.pos
-                with(view) {
-                    with(eventScreenItems[pos] as EventScreenItem.ParticipantItem) {
-                        participantProfiles[participantId]?.let {
-                            setName(if (!it.fullName.isNullOrBlank()) it.fullName!! else if (!it.nickName.isNullOrBlank()) it.nickName!! else if (!it.email.isNullOrBlank()) it.email!! else "#$participantId")
-                            setStatus(if (!it.description.isNullOrBlank()) it.description!! else if (!it.nickName.isNullOrBlank()) it.nickName!! else if (!it.email.isNullOrBlank()) it.email!! else "id пользователя: $participantId")
-                        } ?: run {
-                            setName("[LOADING]")
-                            setStatus("[LOADING]")
-                        }
-                    }
-                }
-            }
-        },
-        override val eventDateItemPresenter: IEventDateItemPresenter =
-            object : IEventDateItemPresenter {
-                private fun isDateExpired(timeInterval: TimeInterval) =
-                    (timeInterval.end < System.currentTimeMillis()) // TODO: отрефакторить: получать время от сущности получаемой через dagger
-
-                override fun onItemClick(view: IEventDateItemView) {
-                    if (!isCurrentUserOwnsEvent || isDateExpired((eventScreenItems[view.pos] as EventScreenItem.EventDateItem).timeInterval)) {
-                        return
-                    }
-                    onEditClick(view)
-                }
-
-                override fun onEditClick(view: IEventDateItemView) {
-                    if (!isCurrentUserOwnsEvent || isDateExpired((eventScreenItems[view.pos] as EventScreenItem.EventDateItem).timeInterval)) {
-                        return
-                    }
-                    openDatePicker((eventScreenItems[view.pos] as EventScreenItem.EventDateItem).timeInterval)
-                }
-
-                override fun onRemoveClick(view: IEventDateItemView) {
-                    val date =
-                        (eventScreenItems[view.pos] as EventScreenItem.EventDateItem).timeInterval
-                    if (!isCurrentUserOwnsEvent || isDateExpired(date)) {
-                        return
-                    }
-                    removeDate(date)
-                }
-
-                override fun bindView(view: IEventDateItemView) = with(view) {
-                    with((eventScreenItems[view.pos] as EventScreenItem.EventDateItem).timeInterval) {
-                        setStartDate(start)
-                        setEndDate(end)
-                        val isDateExpired = isDateExpired(this)
-                        setAsExpired(isDateExpired)
-                        setEditOption(isCurrentUserOwnsEvent && !isDateExpired)
-                    }
-                }
-            },
-        override val noItemsPlaceholderItemPresenter: IItemPresenter<INoItemsPlaceholderItemView> = object :
-            IItemPresenter<INoItemsPlaceholderItemView> {
-            override fun onItemClick(view: INoItemsPlaceholderItemView) {
-                // TODO("Not yet implemented")
-            }
-
-            override fun bindView(view: INoItemsPlaceholderItemView) =
-                with(eventScreenItems[view.pos] as EventScreenItem.NoItemsPlaceholder) {
-                    view.setDescription(description, spanImageRes, spanImagePos)
-                }
-
-        },
-
-        override val textBoxPresenter: IItemPresenter<ITextBoxItemView> = object :
-            IItemPresenter<ITextBoxItemView> {
-            override fun onItemClick(view: ITextBoxItemView) {
-                // TODO("Not yet implemented")
-            }
-
-            override fun bindView(view: ITextBoxItemView) = with(view) {
-                with(eventScreenItems[view.pos] as EventScreenItem.TextBox) {
-                    setValue(value)
-                    setOnValueChangedListener { newValue -> value = newValue }
-                    setEditLock(isEditLocked)
-                    if (hasFocusIntent) {
-                        requestFocus()
-                        hasFocusIntent = false
-                    }
-                }
-            }
-        }
-    ) : IEventScreenListPresenter {
-        override fun getCount() = eventScreenItems.size
-
-        override fun getType(position: Int) = eventScreenItems[position].type.ordinal
-
-        fun setScreenItems(items: List<EventScreenItem>) {
-            eventScreenItems.clear()
-            eventScreenItems.addAll(items)
-            viewState.updateEventScreenList()
-        }
-    }
-
-    val eventScreenListPresenter =
-        EventScreenListPresenter()
-
-    private var isSaveAvailable = true
-
-    private fun saveCallback(successEvent: Event?) {
-        isSaveAvailable = true
-        viewState.enableSaveOptions()
-        if (successEvent == null) {
-            return
-        }
-        eventBeforeEdit = successEvent
-        viewState.hideEditOptions()
-        cancelEdit()
-    }
-
-    fun saveEvent() {
-        // TODO: вынести в кастомную вьюшку
-        if (!isSaveAvailable) {
-            return
-        }
-        isSaveAvailable = false
-        viewState.disableSaveOptions()
-
-        val description =
-            with(eventScreenListPresenter.eventScreenItems.find { item -> item.itemId == EVENT_SCREEN_ITEM_ID.DESCRIPTION_HEADER } as EventScreenItem.FormsHeader<EventScreenItem.ListItem>) {
-                if (items.first() is EventScreenItem.TextBox) {
-                    (items.first() as EventScreenItem.TextBox).value
-                } else {
-                    ""
-                }
-            }
-
-        // TODO: отрефакторить: вынести в extension функцию
-        val dates =
-            with(eventScreenListPresenter.eventScreenItems.find { item -> item.itemId == EVENT_SCREEN_ITEM_ID.DATES_HEADER } as EventScreenItem.FormsHeader<EventScreenItem.ListItem>) {
-                if (items.first() is EventScreenItem.NoItemsPlaceholder) {
-                    TreeSet()
-                } else {
-                    TreeSet<TimeInterval>().apply {
-                        items.forEach {
-                            add((it as EventScreenItem.EventDateItem).timeInterval)
-                        }
-                    }
-                }
-            }
-
-        if (eventBeforeEdit == null) {
-            addEvent(
-                name = (eventScreenListPresenter.eventScreenItems[1] as EventScreenItem.TextForm).value,
-                dates = dates,
-                address = Address(
-                    0.0,
-                    0.0,
-                    (eventScreenListPresenter.eventScreenItems[2] as EventScreenItem.TextForm).value
-                ),
-                description = description,
-                uuid = null,
-                callback = ::saveCallback
-            )
-        } else {
-            val participantsItems =
-                (eventScreenListPresenter.eventScreenItems.find { item -> item.itemId == EVENT_SCREEN_ITEM_ID.PARTICIPANTS_HEADER } as EventScreenItem.FormsHeader<EventScreenItem.ListItem>).items
-            val participantsItemsIterator = participantsItems.iterator()
-            val participantsUserIds =
-                if (participantsItems.first() is EventScreenItem.NoItemsPlaceholder) {
-                    null
-                } else {
-                    LongArray(
-                        participantsItems.size
-                    ) { return@LongArray (participantsItemsIterator.next() as EventScreenItem.ParticipantItem).participantId }
-                }
-
-            val eventAfterEdit = Event(
-                id = eventBeforeEdit!!.id,
-                name = (eventScreenListPresenter.eventScreenItems[1] as EventScreenItem.TextForm).value,
-                ownerUserId = loginRepository.getLocalId()!!,
-                status = Event.Status.ACTUAL,
-                description = description,
-                dates = dates,
-                participantsUserIds = participantsUserIds,
-                city = "PLACEHHOLDER",
-                address = Address(
-                    0.0,
-                    0.0,
-                    (eventScreenListPresenter.eventScreenItems[2] as EventScreenItem.TextForm).value
-                ),
-                mapsFileIds = null,
-                pointsPointIds = null,
-                imageFile = null
-            )
-            editEvent(
-                eventAfterEdit,
-                callback = ::saveCallback
-            )
-        }
-    }
-
-    fun cancelEdit() {
-        eventBeforeEdit?.let {
-            eventScreenListPresenter.setScreenItems(
-                getEventScreenItemsBeforeEdit().apply {
-                    setOfExpandedItems.forEach {
-                        val headerPosition = indexOfFirst { item -> item.itemId == it }
-                        addAll(
-                            headerPosition + 1,
-                            (get(headerPosition) as EventScreenItem.FormsHeader<EventScreenItem.ListItem>).items
-                        )
-                    }
-                })
-            // Log.d("[MYLOG]", "eventScreenListPresenter.eventScreenItems: ${eventScreenListPresenter.eventScreenItems}") // TODO: эта строчка ломает kotlin?
-            eventScreenListPresenter.proeventFormsHeaderItemPresenter.absoluteFormsHeaderPresenter.updateAbsoluteFormsHeader()
-            viewState.updateEventScreenList()
-            viewState.hideEditOptions()
-        } ?: run {
-            localRouter.exit()
-        }
-    }
-
-    private val pickedParticipantsIds: List<Long>
+    val pickedParticipantsIds: List<Long>
         get() {
             val participantsListItems =
                 (eventScreenListPresenter.eventScreenItems.find { item -> item.itemId == EVENT_SCREEN_ITEM_ID.PARTICIPANTS_HEADER } as EventScreenItem.FormsHeader<EventScreenItem.ListItem>).items
@@ -761,10 +170,6 @@ class EventPresenter(localRouter: Router, var eventBeforeEdit: Event?) :
             .show()
     }
 
-    fun openDateEditOptions(timeInterval: TimeInterval) {
-        viewState.showDateEditOptions(pickedDates.indexOf(timeInterval))
-    }
-
     // TODO: отрефакторить: копирует addParticipantsProfiles
     private fun addDateItemView(timeInterval: TimeInterval) {
         // TODO: отрефакторить: перенести это в eventScreenListPresenter
@@ -788,7 +193,7 @@ class EventPresenter(localRouter: Router, var eventBeforeEdit: Event?) :
                         this
                     }
                 }
-            Log.d("[MYLOG]", "listItemPosition($listItemPosition)")
+            // TODO: ВОЗМОЖНЫЙ БАГ: НУЖНО ЛИ ДОБАЛЯТЬ 1?
             val screenItemPosition = headerPosition + 1 + listItemPosition
             items.add(
                 EventScreenItem.EventDateItem(
@@ -807,33 +212,12 @@ class EventPresenter(localRouter: Router, var eventBeforeEdit: Event?) :
             }
             if (!isExpanded) {
                 isExpanded = true
-                setOfExpandedItems.add(EVENT_SCREEN_ITEM_ID.DATES_HEADER)
-                addHeaderItemsToScreenItems(headerPosition, items)
+                eventScreenListPresenter.formsHeaderItemPresenter.setOfExpandedItems.add(EVENT_SCREEN_ITEM_ID.DATES_HEADER)
+                eventScreenListPresenter.eventScreenItems.addAll(headerPosition + 1, items)
             }
         }
         viewState.updateEventScreenList()
         viewState.showEditOptions()
-    }
-
-    fun initParticipantsProfiles(participantsIds: LongArray) {
-//        if (isParticipantsProfilesInitialized) {
-//            return
-//        }
-//        isParticipantsProfilesInitialized = true
-//        for (id in participantsIds) {
-//            profilesRepository.getProfile(id)
-//                .observeOn(uiScheduler)
-//                .subscribe({ profileDto ->
-//                    addParticipantItemView(profileDto!!)
-//                }, {
-//                    val profile = Profile(
-//                        id = id,
-//                        fullName = "Заглушка",
-//                        description = "Профиля нет, или не загрузился",
-//                    )
-//                    addParticipantItemView(profile)
-//                }).disposeOnDestroy()
-//        }
     }
 
     fun initDates(dates: TreeSet<TimeInterval?>) {
@@ -848,10 +232,6 @@ class EventPresenter(localRouter: Router, var eventBeforeEdit: Event?) :
 
     // TODO: отрефакторить: передавать только id
     fun addParticipantsProfiles(participants: Array<Profile>) {
-//        for (participant in participants) {
-//            addParticipantItemView(participant)
-//        }
-
         // TODO: отрефакторить: перенести это в eventScreenListPresenter
         val headerPosition =
             eventScreenListPresenter.eventScreenItems.indexOfFirst { item -> item.itemId == EVENT_SCREEN_ITEM_ID.PARTICIPANTS_HEADER }
@@ -865,9 +245,8 @@ class EventPresenter(localRouter: Router, var eventBeforeEdit: Event?) :
                     ).clear()
                 }
             }
-            val prevLastItemPosition = headerPosition + 1 + items.size
             for (participantProfile in participants) {
-                participantProfiles[participantProfile.id] = participantProfile
+                eventScreenListPresenter.participantItemPresenter.participantProfiles[participantProfile.id] = participantProfile
                 items.add(
                     EventScreenItem.ParticipantItem(
                         participantId = participantProfile.id,
@@ -875,6 +254,8 @@ class EventPresenter(localRouter: Router, var eventBeforeEdit: Event?) :
                     )
                 )
                 if (isExpanded) {
+                    // TODO: я не понял почему не нужно добавлять + 1 к индексу
+                    val prevLastItemPosition = headerPosition + items.size
                     eventScreenListPresenter.eventScreenItems.add(
                         prevLastItemPosition,
                         EventScreenItem.ParticipantItem(
@@ -886,10 +267,11 @@ class EventPresenter(localRouter: Router, var eventBeforeEdit: Event?) :
             }
             if (!isExpanded) {
                 isExpanded = true
-                setOfExpandedItems.add(EVENT_SCREEN_ITEM_ID.PARTICIPANTS_HEADER)
-                addHeaderItemsToScreenItems(headerPosition, items)
+                eventScreenListPresenter.formsHeaderItemPresenter.setOfExpandedItems.add(EVENT_SCREEN_ITEM_ID.PARTICIPANTS_HEADER)
+                eventScreenListPresenter.eventScreenItems.addAll(headerPosition + 1, items)
             }
         }
+
         viewState.updateEventScreenList()
         viewState.showEditOptions()
     }
@@ -900,84 +282,6 @@ class EventPresenter(localRouter: Router, var eventBeforeEdit: Event?) :
 
     fun addEventDate(timeInterval: TimeInterval) {
         addDateItemView(timeInterval)
-    }
-
-    fun clearDates() {
-        viewState.clearDates()
-        pickedDates.clear()
-        isDatesInitialized = false
-    }
-
-    fun clearParticipants() {
-        viewState.clearParticipants()
-        //pickedParticipantsIds.clear()
-        isParticipantsProfilesInitialized = false
-    }
-
-    fun addEventPlace(address: Address?) {
-        localRouter.navigateTo(screens.addEventPlace(address))
-    }
-
-    fun enableDescriptionEdit() {
-        viewState.enableDescriptionEdit()
-    }
-
-    fun expandDescription() {
-        viewState.expandDescription()
-    }
-
-    fun expandMaps() {
-        viewState.expandMaps()
-    }
-
-    fun expandPoints() {
-        viewState.expandPoints()
-    }
-
-    fun expandParticipants() {
-        viewState.expandParticipants()
-    }
-
-    fun expandDates() {
-        viewState.expandDates()
-    }
-
-    fun hideEditOptions() {
-        viewState.hideEditOptions()
-    }
-
-    fun lockEdit() {
-        viewState.lockEdit()
-    }
-
-    fun showMessage(message: String) {
-        viewState.showMessage(message)
-    }
-
-//    fun showAbsoluteBar(
-//        title: String,
-//        iconResource: Int?,
-//        iconTintResource: Int?,
-//        onCollapseScroll: Int,
-//        onCollapse: () -> Unit,
-//        onEdit: () -> Unit
-//    ) {
-//        viewState.showAbsoluteBar(
-//            title,
-//            iconResource,
-//            iconTintResource,
-//            onCollapseScroll,
-//            onCollapse,
-//            onEdit
-//        )
-//    }
-
-    fun unlockNameEdit() {
-        viewState.unlockNameEdit()
-    }
-
-    fun unlockLocationEdit() {
-        viewState.unlockLocationEdit()
     }
 
     // TODO: отрефакторить: копирует removeDateItem()
@@ -993,7 +297,7 @@ class EventPresenter(localRouter: Router, var eventBeforeEdit: Event?) :
             items.removeIf { item -> (item as EventScreenItem.ParticipantItem).participantId == id }
             if (!isExpanded) {
                 isExpanded = true
-                setOfExpandedItems.add(EVENT_SCREEN_ITEM_ID.DATES_HEADER)
+                eventScreenListPresenter.formsHeaderItemPresenter.setOfExpandedItems.add(EVENT_SCREEN_ITEM_ID.DATES_HEADER)
             } else {
                 eventScreenListPresenter.eventScreenItems.removeAt(headerPosition + 1 + indexOfItemToRemove)
             }
@@ -1011,10 +315,6 @@ class EventPresenter(localRouter: Router, var eventBeforeEdit: Event?) :
 
     fun removeParticipant(id: Long) {
         removeParticipantItem(id)
-        // .toList() используется чтобы передать именно копию pickedParticipantsIds, а не ссылку
-//        viewState.removeParticipant(id, pickedParticipantsIds.toList())
-//        viewState.showEditOptions()
-        // pickedParticipantsIds.remove(id)
     }
 
     private fun removeDateItem(timeInterval: TimeInterval) {
@@ -1029,7 +329,7 @@ class EventPresenter(localRouter: Router, var eventBeforeEdit: Event?) :
             items.removeIf { item -> (item as EventScreenItem.EventDateItem).timeInterval == timeInterval }
             if (!isExpanded) {
                 isExpanded = true
-                setOfExpandedItems.add(EVENT_SCREEN_ITEM_ID.DATES_HEADER)
+                eventScreenListPresenter.formsHeaderItemPresenter.setOfExpandedItems.add(EVENT_SCREEN_ITEM_ID.DATES_HEADER)
             } else {
                 eventScreenListPresenter.eventScreenItems.removeAt(headerPosition + 1 + indexOfItemToRemove)
             }
@@ -1047,26 +347,6 @@ class EventPresenter(localRouter: Router, var eventBeforeEdit: Event?) :
 
     fun removeDate(timeInterval: TimeInterval) {
         removeDateItem(timeInterval)
-//        viewState.removeDate(timeInterval, pickedDates.toList())
-//        pickedDates.remove(timeInterval)
-    }
-
-    fun removeDate(position: Int) {
-        viewState.removeDate(pickedDates[position], pickedDates.toList())
-        pickedDates.removeAt(position)
-    }
-
-    fun editDate(position: Int) {
-        Toast.makeText(
-            ProEventApp.instance,
-            "EventPresenter::editDate call;\npickedDates[position]: ${pickedDates[position]};",
-            Toast.LENGTH_LONG
-        )
-            .show()
-    }
-
-    fun hideDateEditOptions() {
-        viewState.hideDateEditOptions()
     }
 
     private fun loadProfiles() {
@@ -1077,7 +357,7 @@ class EventPresenter(localRouter: Router, var eventBeforeEdit: Event?) :
             profilesRepository.getProfile((participantItem as EventScreenItem.ParticipantItem).participantId)
                 .observeOn(uiScheduler)
                 .subscribe({ profileDto ->
-                    participantProfiles[participantItem.participantId] = profileDto!!
+                    eventScreenListPresenter.participantItemPresenter.participantProfiles[participantItem.participantId] = profileDto!!
                     viewState.updateEventScreenList()
                 }, {
                     val profile = Profile(
@@ -1085,7 +365,7 @@ class EventPresenter(localRouter: Router, var eventBeforeEdit: Event?) :
                         fullName = "[ОШИБКА]",
                         description = "Профиля нет, или не загрузился",
                     )
-                    participantProfiles[participantItem.participantId] = profile
+                    eventScreenListPresenter.participantItemPresenter.participantProfiles[participantItem.participantId] = profile
                     viewState.updateEventScreenList()
                 }).disposeOnDestroy()
         }
@@ -1103,9 +383,7 @@ class EventPresenter(localRouter: Router, var eventBeforeEdit: Event?) :
         loadProfiles()
     }
 
-
     fun getEventActionOptions(): List<Pair<String, () -> Unit>> {
-        Log.d("[MYLOG]", "getEventActionOptions")
         return if (eventBeforeEdit!!.ownerUserId == loginRepository.getLocalId()!!) {
             if (eventBeforeEdit!!.status == Event.Status.ACTUAL) {
                 listOf(
