+++ app/src/main/java/ru/myproevent/ui/presenters/events/event/event_screen_items_presenters/forms_header_item_presenters/IHeaderPresenter.kt
@@ -1,0 +1,6 @@
+package ru.myproevent.ui.presenters.events.event.event_screen_items_presenters.forms_header_item_presenters
+
+// TODO: подумать над названием, возмодно стоит переименовать в presenter или остаивть так?
+interface IHeaderPresenter {
+    fun onEditOptionClick(headerPosition: Int)
+}
